+++ lib/index.js
@@ -1,17 +1,32 @@
-var async = require("async");
-var request = require("request");
-var xml2js = require("xml2js");
-var http = require("http");
+var http        = require("http");
 var querystring = require("querystring");
-var _ = require('lodash');
+
+var async      = require("async");
+var request    = require("request");
+var xml2js     = require("xml2js");
+var _          = require('lodash');
 var GoogleAuth = require("google-auth-library");
+
+var Cell      = require('./Cell')
+var Row       = require('./Row')
+var Worksheet = require('./Worksheet')
+var utils     = require('./utils')
+
+var forceArray        = utils.forceArray
+var xmlSafeColumnName = utils.xmlSafeColumnName
+var xmlSafeValue      = utils.xmlSafeValue
+
 
 var GOOGLE_FEED_URL = "https://spreadsheets.google.com/feeds/";
 var GOOGLE_AUTH_SCOPE = ["https://spreadsheets.google.com/feeds"];
 
+
 // The main class that represents a single sheet
 // this is the main module.exports
-var GooogleSpreadsheet = function( ss_key, auth_id, options ){
+function GooogleSpreadsheet( ss_key, auth_id, options ){
+  if(!(this instanceof GooogleSpreadsheet))
+    return new GooogleSpreadsheet( ss_key, auth_id, options )
+
   var self = this;
   var google_auth = null;
   var visibility = 'public';
@@ -137,7 +152,6 @@ var GooogleSpreadsheet = function( ss_key, auth_id, options ){
             return cb( new Error("Sheet is private. Use authentication or make public. (see https://github.com/theoephraim/node-google-spreadsheet#a-note-on-authentication for details)"));
           }
 
-
           if ( body ){
             xml_parser.parseString(body, function(err, result){
               if ( err ) return cb( err );
@@ -151,7 +165,6 @@ var GooogleSpreadsheet = function( ss_key, auth_id, options ){
       }
     });
   }
-
 
 
   // public API methods
@@ -170,7 +183,7 @@ var GooogleSpreadsheet = function( ss_key, auth_id, options ){
       }
       var worksheets = forceArray(data.entry);
       worksheets.forEach( function( ws_data ) {
-        ss_data.worksheets.push( new SpreadsheetWorksheet( self, ws_data ) );
+        ss_data.worksheets.push( new Worksheet( self, ws_data ) );
       })
       cb( null, ss_data );
     });
@@ -241,7 +254,7 @@ var GooogleSpreadsheet = function( ss_key, auth_id, options ){
       var entries = forceArray( data.entry );
       var i=0;
       entries.forEach( function( row_data ) {
-        rows.push( new SpreadsheetRow( self, row_data, entries_xml[ i++ ] ) );
+        rows.push( new Row( self, row_data, entries_xml[ i++ ] ) );
       });
       cb(null, rows);
     });
@@ -255,6 +268,7 @@ var GooogleSpreadsheet = function( ss_key, auth_id, options ){
       }
     });
     data_xml += '</entry>';
+
     self.makeFeedRequest( ["list", ss_key, worksheet_id], 'POST', data_xml, cb );
   }
 
@@ -277,7 +291,7 @@ var GooogleSpreadsheet = function( ss_key, auth_id, options ){
       }
 
       var cells = forceArray(data.entry).map(function( cell_data ){
-        return new SpreadsheetCell( self, worksheet_id, cell_data );
+        return new Cell( self, worksheet_id, cell_data );
       });
 
       cb( null, cells );
@@ -339,168 +353,5 @@ var GooogleSpreadsheet = function( ss_key, auth_id, options ){
   };
 };
 
-// Classes
-var SpreadsheetWorksheet = function( spreadsheet, data ){
-  var self = this;
-
-  self.url = data.id;
-  self.id = data.id.substring( data.id.lastIndexOf("/") + 1 );
-  self.title = data.title;
-  self.rowCount = data['gs:rowCount'];
-  self.colCount = data['gs:colCount'];
-
-  this.getRows = function( opts, cb ){
-    spreadsheet.getRows( self.id, opts, cb );
-  }
-  this.getCells = function (opts, cb) {
-    spreadsheet.getCells( self.id, opts, cb );
-  }
-  this.addRow = function( data, cb ){
-    spreadsheet.addRow( self.id, data, cb );
-  }
-  this.bulkUpdateCells = function( cells, cb ) {
-    spreadsheet.bulkUpdateCells( self.id, cells, cb );
-  }
-  this.del = function ( cb ){
-    spreadsheet.makeFeedRequest( self.url, 'DELETE', null, cb );
-  }
-}
-
-var SpreadsheetRow = function( spreadsheet, data, xml ){
-  Object.keys(data).forEach(function(key) {
-    var val = data[key];
-
-    if(key.substring(0, 4) === "gsx:") {
-      if(typeof val === 'object' && !Object.keys(val).length) {
-        val = null;
-      }
-
-      if (key === "gsx:") {
-        this.gsx = val;
-      } else {
-        this[key.substring(4)] = val;
-      }
-    } else {
-      if (key === "id") {
-        this[key] = val;
-      } else if (val._) {
-        this[key] = val._;
-      } else if ( key === 'link' ){
-        this._links = {};
-
-        forceArray( val ).forEach( function( link ){
-          var $ = link.$;
-          this._links[$.rel] = $.href;
-        }, this);
-      }
-    }
-  }, this);
-
-  Object.defineProperties(this,
-  {
-    _xml: {value: xml},
-
-    id:           {enumerable: false},
-    'app:edited': {enumerable: false},
-    _links:       {enumerable: false},
-
-    save:
-    {
-      value: function( cb ){
-        /*
-        API for edits is very strict with the XML it accepts
-        So we just do a find replace on the original XML.
-        It's dumb, but I couldnt get any JSON->XML conversion to work reliably
-        */
-
-        var data_xml = this._xml;
-        // probably should make this part more robust?
-        data_xml = data_xml.replace('<entry>', "<entry xmlns='http://www.w3.org/2005/Atom' xmlns:gsx='http://schemas.google.com/spreadsheets/2006/extended'>");
-          Object.keys( this ).forEach( function(key) {
-            if (key.substr(0,1) != '_' && typeof this[key] === 'string'){
-              data_xml = data_xml.replace( new RegExp('<gsx:'+xmlSafeColumnName(key)+">([\\s\\S]*?)</gsx:"+xmlSafeColumnName(key)+'>'), '<gsx:'+xmlSafeColumnName(key)+'>'+ xmlSafeValue(self[key]) +'</gsx:'+xmlSafeColumnName(key)+'>');
-            }
-        });
-        spreadsheet.makeFeedRequest( this._links.edit, 'PUT', data_xml, cb );
-      }
-    },
-
-    del:
-    {
-      value: function( cb ){
-        spreadsheet.makeFeedRequest( this._links.edit, 'DELETE', null, cb );
-      }
-    }
-  })
-}
-
-var SpreadsheetCell = function( spreadsheet, worksheet_id, data ){
-  var self = this;
-
-  self.id = data['id'];
-  self.row = parseInt(data['gs:cell']['$']['row']);
-  self.col = parseInt(data['gs:cell']['$']['col']);
-  self.value = data['gs:cell']['_'];
-  self.numericValue = data['gs:cell']['$']['numericValue'];
-  self.inputValue = data['gs:cell']['$']['inputValue'];
-
-  var _hasFormula = self.inputValue.substr(0,1) === '=';
-
-  self['_links'] = [];
-  links = forceArray( data.link );
-  links.forEach( function( link ){
-    self['_links'][ link['$']['rel'] ] = link['$']['href'];
-  });
-
-  self.getValueForSave = function(){
-    if (_hasFormula){
-      return self.inputValue;
-    } else {
-      return xmlSafeValue(self.value);
-    }
-  }
-
-  self.setValue = function(new_value, cb) {
-    self.value = new_value;
-    self.save(cb);
-  };
-
-  self.save = function(cb) {
-    self._needsSave = false;
-
-    var edit_id = 'https://spreadsheets.google.com/feeds/cells/key/worksheetId/private/full/R'+self.row+'C'+self.col;
-    var data_xml =
-    '<entry><id>'+edit_id+'</id>'+
-    '<link rel="edit" type="application/atom+xml" href="'+edit_id+'"/>'+
-    '<gs:cell row="'+self.row+'" col="'+self.col+'" inputValue="'+self.getValueForSave()+'"/></entry>'
-
-    data_xml = data_xml.replace('<entry>', "<entry xmlns='http://www.w3.org/2005/Atom' xmlns:gs='http://schemas.google.com/spreadsheets/2006'>");
-
-    spreadsheet.makeFeedRequest( self['_links']['edit'], 'PUT', data_xml, cb );
-  }
-
-  self.del = function(cb) {
-    self.setValue('', cb);
-  }
-}
 
 module.exports = GooogleSpreadsheet;
-
-//utils
-var forceArray = function(val) {
-  if ( Array.isArray( val ) ) return val;
-  if ( !val ) return [];
-  return [ val ];
-}
-var xmlSafeValue = function(val){
-  if ( val == null ) return '';
-  return String(val).replace(/&/g, '&amp;')
-      .replace(/</g, '&lt;')
-      .replace(/>/g, '&gt;')
-      .replace(/"/g, '&quot;');
-}
-var xmlSafeColumnName = function(val){
-  if (!val) return '';
-  return String(val).replace(/[\s_]+/g, '')
-      .toLowerCase();
-}
